+++ src/Mapbender/CoreBundle/Resources/public/mapbender-model/source.js
@@ -19,187 +19,293 @@
 
 window.Mapbender = Mapbender || {};
 
-window.Mapbender.LayerGroup = (function () {
-    function LayerGroup(title, parent) {
-        this.title_ = title;
-        this.parent = parent || null;
-        this.children = [];
-        this.siblings = [this];
-    }
+(function () {
+    Mapbender.LayerGroup = class LayerGroup {
+        constructor(title, parent) {
+            /**
+             * @type string
+             * @private
+             */
+            this.title_ = title;
+            /**
+             * @type {null|Mapbender.LayerGroup}
+             */
+            this.parent = parent || null;
+            /**
+             *
+             * @type {Mapbender.LayerGroup[]}
+             */
+            this.children = [];
+            /**
+             * @type {Mapbender.LayerGroup[]}
+             */
+            this.siblings = [this];
+        }
 
-    Object.assign(LayerGroup.prototype, {
-        getTitle: function () {
+        getTitle() {
             return this.title_;
-        },
-        getActive: function () {
-            var active = this.getSelected();
-            var parent = this.parent;
+        }
+
+        /**
+         * returnes whether this layergroup and all its parents are selected
+         * @see getSelected
+         * @returns {Boolean}
+         */
+        getActive() {
+            let active = this.getSelected();
+            let parent = this.parent;
             while (parent && active) {
                 active = active && parent.getSelected();
                 parent = parent.parent;
             }
             return active;
-        },
+        }
+
         /**
+         * returns whether this layer is currently selected.
+         * @see getActive
          * @return Boolean
          * @abstract
          */
-        getSelected: function () {
+        getSelected() {
             throw new Error("Invoked abstract LayerGroup.getSelected");
-        },
-        removeChild: function (child) {
+        }
+
+        getParent() {
+            return this.parent;
+        }
+
+        /**
+         * removes the given child layer from this LayerGroup
+         * @param {Mapbender.LayerGroup} child
+         */
+        removeChild(child) {
             [this.children, this.siblings].forEach(function (list) {
-                var index = list.indexOf(child);
+                const index = list.indexOf(child);
                 if (-1 !== index) {
                     list.splice(index, 1);
                 }
             });
         }
-    });
-    return LayerGroup;
-})();
-
-window.Mapbender.Layerset = (function () {
-    function Layerset(title, id, selected) {
-        Mapbender.LayerGroup.call(this, title, null);
-        this.id = id;
-        this.selected = selected;
     }
 
-    Layerset.prototype = Object.create(Mapbender.LayerGroup.prototype);
-    Object.assign(Layerset.prototype, {
-        constructor: Layerset,
-        getId: function () {
+    /**
+     * The most basic LayerGroup definition that is shown in the layer tree but does not display anything itself
+     */
+    Mapbender.Layerset = class Layerset extends Mapbender.LayerGroup {
+        constructor(title, id, selected) {
+            super(title, null);
+            /**
+             * @type {string|number}
+             */
+            this.id = id;
+            /**
+             * @type Boolean
+             */
+            this.selected = selected;
+        }
+
+        getId() {
             return this.id;
-        },
-        getSelected: function () {
+        }
+
+        getSelected() {
             return this.selected;
-        },
-        setSelected: function (state) {
+        }
+
+        setSelected(state) {
             this.selected = !!state;
-        },
-        getSettings: function () {
+        }
+
+        getSettings() {
             return {
                 selected: this.getSelected()
             };
-        },
-        applySettings: function (settings) {
-            var dirty = settings.selected !== this.selected;
-            this.setSelected(settings.selected);
+        }
+
+        /**
+         * Changes all given layer settings
+         * @see getSettings
+         * @param {{selected?: Boolean}} settings
+         * @returns {boolean} true if at least one attribute has been changed
+         */
+        applySettings(settings) {
+            let dirty = false;
+            if ("selected" in settings) {
+                dirty = settings.selected !== this.selected;
+                this.setSelected(settings.selected);
+            }
             return dirty;
         }
-    });
-    return Layerset;
-})();
-
-window.Mapbender.Source = (function () {
-    function Source(definition) {
-        Mapbender.LayerGroup.call(this, definition.title, null);
-        if (definition.id || definition.id === 0) {
-            this.id = '' + definition.id;
-        }
-        this.type = definition.type;
-        this.configuration = definition.configuration;
-        this.wmsloader = definition.wmsloader;
-        var sourceArg = this;
-        this.configuration.children = (this.configuration.children || []).map(function (childDef) {
-            return Mapbender.SourceLayer.factory(childDef, sourceArg, null)
-        });
-        this.children = this.configuration.children;
-        this.configuredSettings_ = this.getSettings();
     }
 
-    Source.typeMap = {};
     /**
-     * @param {*} definition
-     * @returns {Mapbender.Source}
+     * A source represents a data source that is displayed on the map. WmsSource, WmtsSource, GeoJsonSource etc. extend
+     * from this base class. A source can have one or more SourceLayers.
+     * They must be defined in the constructor using definition.configuration.children
+     * @see SourceLayer
+     * @abstract
      */
-    Source.factory = function (definition) {
-        var typeClass = Source.typeMap[definition.type];
-        if (!typeClass) {
-            typeClass = Source;
+    Mapbender.Source = class Source extends Mapbender.LayerGroup {
+        constructor(definition) {
+            super(definition.title, null);
+
+            /**
+             * A unique identifier for this source
+             * @type {null|string}
+             */
+            this.id = null;
+            if (definition.id || definition.id === 0) {
+                this.id = '' + definition.id;
+            }
+
+            /**
+             * The native OpenLayers layers
+             * @see createNativeLayers
+             * @see getNativeLayers
+             * @type {ol.Layer[]}
+             */
+            this.nativeLayers = [];
+
+            /**
+             * Indicates whether this source has been added by the user during runtime, e.g. by using the WMSLoader
+             * dynamic sources are removed when the view is reset
+             * @type {boolean}
+             */
+            this.isDynamicSource = false;
+            if ("isDynamicSource" in definition) this.isDynamicSource = definition.isDynamicSource;
+
+            /**
+             * a unique identifier for the type of source, e.g. 'wms' or 'geojson'
+             * @type {string}
+             */
+            this.type = definition.type || 'undefined';
+
+            /**
+             * Configuration options for this source, keys depend on the source type
+             * @type {object}
+             */
+            this.configuration = definition.configuration || {};
+            this.configuration.options = this.configuration.options || {};
+            this.configuration.children = (this.configuration.children || []).map(childDef => Mapbender.SourceLayer.factory(childDef, this, null));
+
+            this.children = this.configuration.children;
+            this.configuredSettings_ = this.getSettings();
+
+            /**
+             * the (optional) layerset this source is assigned to
+             * @type {null|Mapbender.Layerset}
+             */
+            this.layerset = null;
         }
-        return new typeClass(definition);
-    };
-    Source.prototype = Object.create(Mapbender.LayerGroup.prototype);
-    Object.assign(Source.prototype, {
-        constructor: Source,
+
         /**
+         * Creates a new source instance
+         * @param {*} definition object containing at least the attribute 'type'
+         * @returns {Mapbender.Source}
+         */
+        static factory(definition) {
+            var typeClass = Source.typeMap[definition.type];
+            if (!typeClass) {
+                typeClass = Mapbender.Source;
+            }
+            return new typeClass(definition);
+        }
+
+        /**
+         * Creates the native OpenLayers layers required for this source and stores them in the class variable nativeLayers
+         * @abstract
          * @param {String} srsName
          * @param {Object} [mapOptions]
-         * @return {Array<Object>}
+         * @return {ol.Layer[]}
          */
-        createNativeLayers: function (srsName, mapOptions) {
+        createNativeLayers(srsName, mapOptions) {
             console.error("Layer creation not implemented", this);
             throw new Error("Layer creation not implemented");
-        },
+        }
+
         /**
-         * @param {String} srsName
-         * @param {Object} [mapOptions]
-         * @return {Array<Object>}
+         * Layer state has changed, e.g. layers have been reordered, deselected etc. Refresh the view in the native layers.
+         * @abstract
          */
-        initializeLayers: function (srsName, mapOptions) {
-            this.nativeLayers = this.createNativeLayers(srsName, mapOptions);
-            return this.nativeLayers;
-        },
-        getActive: function () {
-            var upstream = Mapbender.LayerGroup.prototype.getActive.call(this);
-            // NOTE: (only) WmsLoader sources don't have a layerset
+        updateEngine() {
+            console.error("Update engine not implemented", this);
+            throw new Error("Update engine not implemented");
+        }
+
+        /**
+         * @returns {ol.Layer[]}
+         */
+        getNativeLayers() {
+            return this.nativeLayers.slice();
+        }
+
+        getActive() {
+            const upstream = super.getActive();
+            // NOTE: some sources like e.g. WmsLoader sources don't have a layerset
             return upstream && (!this.layerset || this.layerset.getSelected());
-        },
-        id: null,
-        title: null,
-        type: null,
-        configuration: {},
-        nativeLayers: [],
-        recreateOnSrsSwitch: false,
-        wmsloader: false,
-        destroyLayers: function (olMap) {
+        }
+
+        /**
+         * destroys all currently assigned native layers
+         * @param {ol.Map} olMap
+         */
+        destroyLayers(olMap) {
             if (this.nativeLayers && this.nativeLayers.length) {
                 this.nativeLayers.map(function (olLayer) {
                     Mapbender.mapEngine.destroyLayer(olMap, olLayer);
                 });
             }
             this.nativeLayers = [];
-        },
-        getNativeLayers: function () {
-            return this.nativeLayers.slice();
-        },
-        getSettings: function () {
+        }
+
+        getSettings() {
             return {
                 opacity: this.configuration.options.opacity
             };
-        },
-        getConfiguredSettings: function () {
+        }
+
+        getConfiguredSettings() {
             return Object.assign({}, this.configuredSettings_);
-        },
+        }
+
+        /**
+         * Returns all layers that support feature info
+         * @return {Array<Mapbender.SourceLayer>}
+         */
+        getFeatureInfoLayers() {
+            return [];
+        }
+
         /**
          * @param {SourceSettings} settings
-         * @return {boolean}
+         * @return {boolean} true if at least one setting has been changed
          */
-        applySettings: function (settings) {
-            var diff = this.diffSettings(this.getSettings(), settings);
+        applySettings(settings) {
+            const diff = this.diffSettings(this.getSettings(), settings);
             if (diff) {
                 this.applySettingsDiff(diff);
                 return true;
             } else {
                 return false;
             }
-        },
+        }
+
         /**
          * @param {SourceSettingsDiff} diff
          */
-        applySettingsDiff: function (diff) {
+        applySettingsDiff(diff) {
             if (diff && typeof (diff.opacity) !== 'undefined') {
                 this.setOpacity(diff.opacity);
             }
-        },
+        }
+
         /**
          * @param {SourceSettings} from
          * @param {SourceSettings} to
          * @return {SourceSettingsDiff|null}
          */
-        diffSettings: function (from, to) {
+        diffSettings(from, to) {
             // before v4, only the selectedIds were saved as an array.
             // since v4, an object with id and name is saved are saved encapsulated as an array of objects to
             // check for selectedIds for backwards compatibility
@@ -242,13 +348,14 @@ window.Mapbender.Source = (function () {
             }
             // null if completely empty
             return Object.keys(diff).length && diff || null;
-        },
+        }
+
         /**
          * @param {SourceSettings} base
          * @param {SourceSettingsDiff} diff
          * @return {SourceSettings}
          */
-        mergeSettings: function (base, diff) {
+        mergeSettings(base, diff) {
             var settings = Object.assign({}, base);
             if (typeof (diff.opacity) !== 'undefined') {
                 settings.opacity = diff.opacity;
@@ -258,25 +365,39 @@ window.Mapbender.Source = (function () {
             });
             settings.selectedLayers = settings.selectedLayers.concat((diff || {}).activate || []);
             return settings;
-        },
+        }
 
-        checkRecreateOnSrsSwitch: function (oldProj, newProj) {
-            return this.recreateOnSrsSwitch;
-        },
-        getNativeLayer: function (index) {
-            var layer = this.nativeLayers[index || 0] || null;
-            var c = this.nativeLayers.length;
+        /**
+         * indicates whether this source should be recreated when a srs change occurs
+         * @param {string} oldProj
+         * @param {string} newProj
+         * @returns {boolean}
+         */
+        checkRecreateOnSrsSwitch(oldProj, newProj) {
+            return false;
+        }
+
+        /**
+         * Gets the native OpenLayers layer of the specified index
+         * @see createNativeLayers
+         * @param {number|undefined} index
+         * @returns {ol.Layer|null}
+         */
+        getNativeLayer(index) {
+            const layer = this.nativeLayers[index || 0] || null;
+            const c = this.nativeLayers.length;
             if (typeof index === 'undefined' && c !== 1) {
                 console.warn("Mapbender.Source.getNativeLayer called on a source with flexible layer count; currently " + c + " native layers");
             }
             return layer;
-        },
+        }
+
         /**
          * @param {string} id
          * @return {SourceLayer}
          */
-        getLayerById: function (id) {
-            var foundLayer = null;
+        getLayerById(id) {
+            let foundLayer = null;
             Mapbender.Util.SourceTree.iterateLayers(this, false, function (sourceLayer) {
                 if ((sourceLayer.options?.id ?? sourceLayer.id) === id) {
                     foundLayer = sourceLayer;
@@ -285,58 +406,79 @@ window.Mapbender.Source = (function () {
                 }
             });
             return foundLayer;
-        },
-        getRootLayer: function () {
+        }
+
+        /**
+         * @returns {Mapbender.SourceLayer}
+         */
+        getRootLayer() {
             return this.configuration.children[0];
-        },
-        _reduceBboxMap: function (bboxMap, projCode) {
-            if (bboxMap && Object.keys(bboxMap).length) {
-                if (projCode) {
-                    if (bboxMap[projCode]) {
-                        var reduced = {};
-                        reduced[projCode] = bboxMap[projCode];
-                        return reduced;
-                    }
-                    return null;
-                }
-                return bboxMap;
-            }
-            return null;
-        },
-        getLayerBounds: function (layerId, projCode, inheritFromParent) {
-            var layer;
-            if (layerId) {
-                layer = this.getLayerById(layerId);
-            } else {
-                // root layer
-                layer = this.configuration.children[0];
-            }
+        }
+
+        /**
+         * @param {number|undefined} layerId
+         * @param {string} projCode
+         * @param {boolean} inheritFromParent
+         * @returns {number[]|boolean} false if bounds could not be calculated
+         */
+        getLayerBounds(layerId, projCode, inheritFromParent) {
+            const layer = layerId ? this.getLayerById(layerId) : this.configuration.children[0];
+
             if (!layer) {
                 console.warn("No layer, unable to calculate bounds");
                 return false;
             }
             return layer.getBounds(projCode, inheritFromParent) || null;
-        },
-        setOpacity: function (value) {
+        }
+
+        /**
+         * @param {number} value between 0 and 1
+         */
+        setOpacity(value) {
             this.configuration.options.opacity = value;
             this.nativeLayers.map(function (layer) {
                 layer.setOpacity(value);
             });
-        },
-        _bboxArrayToBounds: function (bboxArray, projCode) {
+        }
+
+        /**
+         * reorders the layers of this source. The default implementation forwards to
+         * [Mapbender.Geo.SourceHandler.setLayerOrder] which works for hierarchical sources like WMS and WMTS
+         * @param {string[]} newLayerIdOrder the layer ids in their new order
+         */
+        setLayerOrder(newLayerIdOrder) {
+            Mapbender.Geo.SourceHandler.setLayerOrder(this, newLayerIdOrder);
+        }
+
+        _bboxArrayToBounds(bboxArray, projCode) {
             return Mapbender.mapEngine.boundsFromArray(bboxArray);
-        },
-        _getPrintBaseOptions: function () {
+        }
+
+        _getPrintBaseOptions() {
             return {
                 type: this.configuration.type,
                 sourceId: this.id,
-                // @todo: use live native layer opacity?
                 opacity: this.configuration.options.opacity
             };
-        },
-        // Custom toJSON for mbMap.getMapState()
-        // Drops nativeLayers to avoid circular references
-        toJSON: function () {
+        }
+
+        /**
+         * Returns information that is passed to the printing service when printing or exporting a map
+         * @param {Number[]} bounds
+         * @param {Number} scale
+         * @param {String} srsName
+         * @return {Array<Object>}
+         */
+        getPrintConfigs(bounds, scale, srsName) {
+            return [];
+        }
+
+        /**
+         * Custom toJSON for mbMap.getMapState()
+         * Drops nativeLayers to avoid circular references
+         * @returns {{configuration: Object, id: (string|null), title, type: string}}
+         */
+        toJSON() {
             return {
                 id: this.id,
                 title: this.title,
@@ -344,34 +486,96 @@ window.Mapbender.Source = (function () {
                 configuration: this.configuration
             };
         }
-    });
-    return Source;
-}());
-
-window.Mapbender.SourceLayer = (function () {
-    function SourceLayer(definition, source, parent) {
-        Mapbender.LayerGroup.call(this, ((definition || {}).options || {}).title || '', parent)
-        this.options = definition.options || {};
-        this.state = definition.state || {};
-        this.source = source;
-        var childDefs = definition.children || [];
-        var i, child, childDef;
-        for (i = 0; i < childDefs.length; ++i) {
-            childDef = childDefs[i];
-            child = SourceLayer.factory(childDef, source, this);
-            child.siblings = this.children;
-            this.children.push(child);
-        }
-        this.siblings = [this];
     }
 
-    SourceLayer.prototype = Object.create(Mapbender.LayerGroup.prototype);
-    Object.assign(SourceLayer.prototype, {
-        constructor: SourceLayer,
-        // need custom toJSON for getMapState call
-        toJSON: function () {
+
+    /**
+     * @abstract
+     */
+    Mapbender.SourceLayer = class SourceLayer extends Mapbender.LayerGroup {
+        constructor(definition, source, parent) {
+            super(((definition || {}).options || {}).title || '', parent)
+            this.options = definition.options || {};
+            this.options.treeOptions = this.options.treeOptions || {
+                selected: true,
+                info: false,
+                toggle: true,
+                allow: {selected: true, info: false, toggle: true,}
+            };
+
+            this.state = definition.state || {
+                info: null,
+                outOfBounds: false,
+                outOfScale: false,
+                visibility: true,
+            };
+
+            this.source = source;
+            var childDefs = definition.children || [];
+            var i, child, childDef;
+            for (i = 0; i < childDefs.length; ++i) {
+                childDef = childDefs[i];
+                child = Mapbender.SourceLayer.factory(childDef, source, this);
+                child.siblings = this.children;
+                this.children.push(child);
+            }
+            this.siblings = [this];
+        }
+
+        static factory(definition, source, parent) {
+            let typeClass = SourceLayer.typeMap[source.type];
+            if (!typeClass) {
+                typeClass = Mapbender.SourceLayer;
+            }
+            return new typeClass(definition, source, parent);
+        }
+
+        /**
+         * is this layer selected in the layertree
+         * Caution: This does not mean it's visible, parent layers might be unselected
+         * @returns {boolean}
+         */
+        getSelected() {
+            return this.options.treeOptions.selected;
+        }
+
+        setSelected(state) {
+            this.options.treeOptions.selected = !!state;
+        }
+
+        getId() {
+            return this.options.id;
+        }
+
+        getName() {
+            return this.options.name;
+        }
+
+        /**
+         * Should the layer be displayed at this scale level?
+         * @param {number} scale
+         * @returns {boolean}
+         */
+        isInScale(scale) {
+            return true;
+        }
+
+        /**
+         * Does the layer has contents in this extent?
+         * @param {number[]} extent
+         * @param {string} srsName
+         * @returns {boolean}
+         */
+        intersectsExtent(extent, srsName) {
+            return true;
+        }
+
+        /**
+         * need custom toJSON for getMapState call
+         */
+        toJSON() {
             // Skip the circular-ref inducing properties 'siblings', 'parent' and 'source'
-            var r = {
+            const r = {
                 options: this.options,
                 state: this.state
             };
@@ -379,24 +583,41 @@ window.Mapbender.SourceLayer = (function () {
                 r.children = this.children;
             }
             return r;
-        },
-        getParent: function () {
-            return this.parent;
-        },
-        remove: function () {
-            var index = this.siblings.indexOf(this);
-            if (index !== -1) {
-                this.siblings.splice(index, 1);
-                if (!this.siblings.length && this.parent) {
-                    return this.parent.remove();
-                } else {
-                    return this.options.id;
-                }
-            } else {
+        }
+
+        /**
+         * removes this layer from the source tree
+         * @returns {string|null} the if of the removed layer or null of the layer had no parent
+         */
+        remove() {
+            const index = this.siblings.indexOf(this);
+            if (index === -1) {
                 return null;
             }
-        },
-        getBounds: function (projCode, inheritFromParent) {
+
+            this.siblings.splice(index, 1);
+            if (!this.siblings.length && this.parent && this.parent.remove) {
+                return this.parent.remove();
+            }
+
+            return this.options.id;
+        }
+
+        /**
+         * @param {Mapbender.SourceLayer} child
+         */
+        addChild(child) {
+            this.children.push(child);
+            this.children.forEach((child) => child.siblings = this.children);
+            Mapbender.Model.updateSource(this.source);
+        }
+
+        /**
+         * @param {string} projCode
+         * @param {boolean} inheritFromParent
+         * @returns {number[]|boolean} false if bounds could not be calculated
+         */
+        getBounds(projCode, inheritFromParent) {
             var bboxMap = this.options.bbox;
             var srsOrder = [projCode].concat(Object.keys(bboxMap));
             for (var i = 0; i < srsOrder.length; ++i) {
@@ -412,8 +633,23 @@ window.Mapbender.SourceLayer = (function () {
                 return this.parent.getBounds(projCode, true);
             }
             return null;
-        },
-        hasBounds: function () {
+        }
+
+        /**
+         * Returns the legend for this layer. The legend can be either an external
+         * url (e.g. for WMS services) or a style definition that is rendered on a canvas
+         *
+         * @return {null|{type: 'url', url: string, topLevel: boolean}|LegendDefinition}
+         */
+        getLegend() {
+            return null;
+        }
+
+        /**
+         * is this layer restricted to spatial bbox?
+         * @returns {boolean}
+         */
+        hasBounds() {
             var layer = this;
             do {
                 if (Object.keys(layer.options.bbox).length) {
@@ -423,14 +659,39 @@ window.Mapbender.SourceLayer = (function () {
             } while (layer);
             return false;
         }
-    });
-    SourceLayer.typeMap = {};
-    SourceLayer.factory = function (definition, source, parent) {
-        var typeClass = SourceLayer.typeMap[source.type];
-        if (!typeClass) {
-            typeClass = SourceLayer;
+
+        /**
+         * Returns a list of menu options supported by this layer.
+         * Core mapbender menu options:
+         * - layerremove: Deletes this layer
+         * - metadata: Opens metadata in a new window. options.medataUrl should be defined
+         * - opacity: Opacity slider between 0 and 1. See {Mapbender.Source.setOpacity}
+         * - dimension: selection slider for dimensions like e.g. time
+         * - zoomtolayer: Changes the map's view to fit the layer
+         * @returns {string[]}
+         */
+        getSupportedMenuOptions() {
+            const supported = ['layerremove'];
+            if (this.options.metadataUrl) {
+                supported.push('metadata');
+            }
+            // opacity + dimension are only available on root layer
+            if (!this.getParent()) {
+                supported.push('opacity');
+                if ((this.source.configuration.options.dimensions || []).length) {
+                    supported.push('dimension');
+                }
+            }
+            if (this.hasBounds()) {
+                supported.push('zoomtolayer');
+            }
+            if (this.options.availableStyles && this.options.availableStyles.length > 1) {
+                supported.push('select_style');
+            }
+            return supported;
         }
-        return new typeClass(definition, source, parent);
-    };
-    return SourceLayer;
+    }
+
+    Mapbender.Source.typeMap = {};
+    Mapbender.SourceLayer.typeMap = {};
 }());