+++ src/Mapbender/CoreBundle/Resources/public/mapbender.element.overview.js
@@ -139,7 +139,7 @@
                 // Legacy HACK: Overview ignores backend settings on instance layers, enables all children
                 //        of the root layer with non-empty names, ignores every other layer
                 if (source.type !== 'wms' || source.hasVisibleLayers(srsName)) {
-                    source.initializeLayers(srsName);
+                    source.createNativeLayers(srsName);
                     layers = layers.concat(source.getNativeLayers());
                 }
             }